+++ test/jest-component/features/steps/reusable-steps.ts
@@ -4,12 +4,9 @@ import * as request from 'supertest';
 import Ajv from 'ajv';
 import addFormats from 'ajv-formats';
 import {
-  app, schema, accountRepository,
-  tokensRepository, generateTokenPair, context
+  app, schema, context
 } from '../../steps-config';
-import { USER_DB, PASSWORD } from '../../../mocks/db';
 
-const TTL = parseInt(process.env.ACCESS_TOKEN_TTL);
 
 
 export const reusableSteps: StepDefinitions = ({ and, when, then }) => {
@@ -39,10 +36,11 @@ export const reusableSteps: StepDefinitions = ({ and, when, then }) => {
     }
     expect(valid).toBe(true);
   });
-  then(
-    /^error response code (\d+) returned, description: (.*)$/,
-    (error, p) => {});
-  then(/^error response code (\d+) returned$/, (error) => {
+  then(/^error response code (\d+) returned()$/, (error) => {});
+//then(
+//    /^error response code (\d+) returned, description: (.*)$/,
+//    (error, p) => {});
+  then(/^error response code (\d+) returned, description: (.*)$/, (error, description) => {
     if (error == 500) {
       expect(context.response.status).toBe(HttpStatus.INTERNAL_SERVER_ERROR);
       expect(context.response.body).toEqual({
@@ -54,9 +52,16 @@ export const reusableSteps: StepDefinitions = ({ and, when, then }) => {
       });
     } else if (error == 400) {
       expect(context.response.status).toBe(HttpStatus.BAD_REQUEST);
-      expect(context.response.body.errors[0]).toBe(
-        'eventName should not be empty',
-      );
+      if (context.path == '/v1/events') {
+        expect(context.response.body.errors[0]).toBe('eventName should not be empty');    
+      } else if ((context.path == '/v1/account/login') && (description == "without pass")) {
+        expect(context.response.body.errors[0]).toBe('password should not be empty');
+      } else if (context.path == '/v1/account/login') {
+        expect(context.response.body.errors[0]).toBe('email should not be empty');
+      }
+    } else if (error == 401) {
+      expect(context.response.status).toBe(HttpStatus.UNAUTHORIZED);
+      expect(context.response.body.errors[0]).toBe('email password combination is not correct')
     } else {
       expect(true).toBe(false);
     }
@@ -69,15 +74,7 @@ export const reusableSteps: StepDefinitions = ({ and, when, then }) => {
     } else if (code_n == 200) {
       console.log(context.response);
       expect(context.response.status).toBe(HttpStatus.OK);
-      expect(context.response.body).toMatchObject({
-        accessToken: 'ACCESS_TOKEN',
-        refreshToken: 'REFRESH_TOKEN',
-        ttl: TTL,
-        profile: {
-          name: USER_DB.name,
-          email: USER_DB.email,
-        },
-      });
+      
     } else {
       expect(false).toBe(true);
     }
